+++ examples/canopen/can_monitoring.py
@@ -1,4 +1,4 @@
-import sys
+import argparse
 from typing import List
 
 import numpy as np
@@ -6,23 +6,27 @@ from numpy.typing import NDArray
 
 from ingenialink.canopen.network import CAN_BAUDRATE, CAN_DEVICE, CanopenNetwork
 from ingenialink.canopen.register import CanopenRegister
+from ingenialink.exceptions import ILRegisterNotFoundError
 
 
-def monitoring_example() -> List[NDArray[np.float_]]:
+def monitoring_example(args: argparse.Namespace) -> List[NDArray[np.float_]]:
     registers_key = [
         "DRV_PROT_TEMP_VALUE",
     ]
 
-    net = CanopenNetwork(device=CAN_DEVICE.IXXAT,
-                         channel=0,
-                         baudrate=CAN_BAUDRATE.Baudrate_1M)
-    nodes = net.scan_slaves()
-    servo = net.connect_to_slave(
-        target=nodes[0],
-        dictionary='../../resources/dictionaries/eve-net-c_can_1.8.1.xdf')
+    can_device = CAN_DEVICE(args.transceiver)
+    can_baudrate = CAN_BAUDRATE(args.baudrate)
+    net = CanopenNetwork(device=can_device, channel=args.channel, baudrate=can_baudrate)
+    servo = net.connect_to_slave(target=args.node_id, dictionary=args.dictionary_path)
     # Monitoring
     # Remove all mapped registers
-    servo.monitoring_disable()
+    try:
+        servo.monitoring_disable()
+    except ILRegisterNotFoundError:
+        print("Monitoring is not available for this drive")
+        net.disconnect_from_slave(servo)
+        return []
+
     servo.monitoring_remove_all_mapped_registers()
 
     # Calculate the monitoring frequency
@@ -49,13 +53,13 @@ def monitoring_example() -> List[NDArray[np.float_]]:
         tmp_mon_data.append([])
         monitor_data.append([])
     # Configure monitoring SOC as forced
-    servo.write('MON_CFG_SOC_TYPE', 0, subnode=0)
+    servo.write("MON_CFG_SOC_TYPE", 0, subnode=0)
     # Configure monitoring EoC as number of samples
-    servo.write('MON_CFG_EOC_TYPE', 3, subnode=0)
+    servo.write("MON_CFG_EOC_TYPE", 3, subnode=0)
     # Configure number of samples
     window_samples = 599
     total_num_samples = window_samples
-    servo.write('MON_CFG_WINDOW_SAMP', window_samples, subnode=0)
+    servo.write("MON_CFG_WINDOW_SAMP", window_samples, subnode=0)
     # Enable monitoring
     servo.monitoring_enable()
     # Check monitoring status
@@ -65,7 +69,7 @@ def monitoring_example() -> List[NDArray[np.float_]]:
     if (monitor_status & 0x1) != 1:
         raise ValueError(f"ERROR MONITOR STATUS: {monitor_status}")
     # Force Trigger
-    servo.write('MON_CMD_FORCE_TRIGGER', 1, subnode=0)
+    servo.write("MON_CMD_FORCE_TRIGGER", 1, subnode=0)
     sampling_time_s = 1 / sampling_freq
     data_obtained = False
 
@@ -82,23 +86,14 @@ def monitoring_example() -> List[NDArray[np.float_]]:
                 servo.monitoring_read_data()
                 for idx, key in enumerate(registers_key):
                     index = idx
-                    tmp_monitor_data = servo. \
-                        monitoring_channel_data(index)
-                    tmp_mon_data[index] = \
-                        tmp_mon_data[index] + tmp_monitor_data
+                    tmp_monitor_data = servo.monitoring_channel_data(index)
+                    tmp_mon_data[index] = tmp_mon_data[index] + tmp_monitor_data
                     if len(tmp_mon_data[index]) >= total_num_samples:
-                        tmp_mon_data[index] = np.resize(
-                            tmp_mon_data[index], total_num_samples
-                        )
-                        data_x = np.arange(
-                            (window_samples) * sampling_time_s,
-                            sampling_time_s
-                        )
+                        tmp_mon_data[index] = np.resize(tmp_mon_data[index], total_num_samples)
+                        data_x = np.arange((window_samples) * sampling_time_s, sampling_time_s)
 
                         if data_x.size > len(tmp_mon_data[index]):
-                            data_x = np.resize(
-                                data_x, len(tmp_mon_data[index])
-                            )
+                            data_x = np.resize(data_x, len(tmp_mon_data[index]))
                         data_y = tmp_mon_data[index]
                         monitor_data[index] = np.round(data_y, decimals=2)
                         data_obtained = True
@@ -106,7 +101,7 @@ def monitoring_example() -> List[NDArray[np.float_]]:
                     # Single-shot mode
                     read_process_finished = True
         except Exception as e:
-            print('Exception monitoring: {}'.format(e))
+            print("Exception monitoring: {}".format(e))
             break
     print("Finished")
 
@@ -114,6 +109,29 @@ def monitoring_example() -> List[NDArray[np.float_]]:
     return monitor_data
 
 
-if __name__ == '__main__':
-    monitoring_example()
-    sys.exit(0)
+def setup_command() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(description="Canopen example")
+    parser.add_argument("-d", "--dictionary_path", help="Path to drive dictionary", required=True)
+    parser.add_argument("-n", "--node_id", default=32, type=int, help="Node ID")
+    parser.add_argument(
+        "-t",
+        "--transceiver",
+        default="ixxat",
+        choices=["pcan", "kvaser", "ixxat"],
+        help="CAN transceiver",
+    )
+    parser.add_argument(
+        "-b",
+        "--baudrate",
+        default=1000000,
+        type=int,
+        choices=[50000, 100000, 125000, 250000, 500000, 1000000],
+        help="CAN baudrate",
+    )
+    parser.add_argument("-c", "--channel", default=0, type=int, help="CAN transceiver channel")
+    return parser.parse_args()
+
+
+if __name__ == "__main__":
+    args = setup_command()
+    monitoring_example(args)