+++ examples/ethercat/ecat_load_firmware.py
@@ -1,21 +1,29 @@
-import sys
+import argparse
 
 from ingenialink.ethercat.network import EthercatNetwork
 
 
-def ecat_load_fw() -> None:
-    net = EthercatNetwork("\\Device\\NPF_{192D1D2F-C684-467D-A637-EC07BD434A63}")
+def ecat_load_fw(args: argparse.Namespace) -> None:
+    net = EthercatNetwork(args.interface)
+    boot_in_app = args.firmware_path.endswith(".sfu")
+    net.load_firmware(args.firmware_path, boot_in_app, slave_id=args.slave_id)
 
-    # Set boot_in_app to True if the firmware file is of type .sfu
-    # It is important to set this parameter carefully as an incorrect configuration can cause the
-    # drive to become non-functional!
-    boot_in_app = False
 
-    net.load_firmware(
-        fw_file="../../resources/firmware/cap-net-e_0.7.1.lfu", boot_in_app=boot_in_app
+def setup_command() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(description="EtherCAT connection example script.")
+    interface_help = """Network adapter interface name. To find it: \n
+    - On Windows, \\Device\\NPF_{id}. To get the id, run the command: wmic nic get name, guid \n
+    - On linux, run the command: ip link show
+    """
+    parser.add_argument("-i", "--interface", type=str, help=interface_help, required=True)
+    parser.add_argument(
+        "-f", "--firmware_path", type=str, help="Path to the firmware file.", required=True
     )
+    parser.add_argument("-s", "--slave_id", type=int, help="Slave ID.", default=1)
+    args = parser.parse_args()
+    return args
 
 
 if __name__ == "__main__":
-    ecat_load_fw()
-    sys.exit(0)
+    args = setup_command()
+    ecat_load_fw(args)